renames NoTokensCard to NoTokensHeldCard
